test/requirements.txt
# Daita Agents Framework
daita-agents==0.12.1

# LLM provider
openai>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
